Harden docs search and telemetry
