+++ docs/scripts/browser/exceptionless-client.ts
@@ -6,6 +6,7 @@ const apiKey = options.get("apiKey")
 if (apiKey) {
   await Exceptionless.startup((clientConfig) => {
     clientConfig.apiKey = apiKey
+    clientConfig.includePrivateInformation = false
     clientConfig.defaultTags.push("Website")
 
     const serverUrl = options.get("serverUrl")
@@ -18,11 +19,10 @@ if (apiKey) {
       clientConfig.version = version
     }
 
-    clientConfig.useSessions()
     clientConfig.addPlugin("site-context", 10, async (context) => {
       context.event.data = context.event.data || {}
       context.event.data["@page"] = location.pathname
-      context.event.data["@page.url"] = location.href
+      context.event.data["@page.url"] = `${location.origin}${location.pathname}`
       context.event.data["@page.title"] = document.title
     })
   })
