+++ src/Exceptionless.Web/ClientApp/.npmrc
@@ -1,2 +1,2 @@
 engine-strict=true
-min-release-age=3
+min-release-age=7
